Throttle bug was solved / Throttle range changed to [0,100] (int). Friciton decel. applied. Normal brakes maped to JS' Triangle button. RefreshVerticalSpeed() currently being rewritten from scratch as RefreshAttitude().

+++ Aero390Spoilers/Ownship/Ownship.cs
@@ -203,7 +203,7 @@ namespace Ownship
         public double GrossWeightLbs { get; set; }
         public double BaroSettingmmHg { get; set; }
         public double AltitudeASL { get; set; }
-        public int IasKts { get; set; }
+        public double IasKts { get; set; }
         public int IasOverspeed { get; set; }
         public int IasNES { get; set; }
         public int IasStall { get; set; }
